requirements.txt
Flask==3.1.0
flask-cors==6.0.1
python-dotenv==1.1.0
pandas==2.2.3
langchain==0.1.16
langchain-community==0.0.33
ollama==0.6.1
qdrant-client==1.15.1